assignment_1/another-producer/requirements.txt
kafka-python
mimesis
